change display to flex for body containers
